rename proxy backend packages to be of the form fooproxy
This makes a clear distinction between the disk cache package and the
proxy backend packages, and avoids http_test.go ambiguity in go test
error logs (we previously had two http_test.go files).
