+++ BUILD.bazel
@@ -21,9 +21,9 @@ go_library(
     deps = [
         "//cache:go_default_library",
         "//cache/disk:go_default_library",
-        "//cache/gcs:go_default_library",
-        "//cache/http:go_default_library",
-        "//cache/s3:go_default_library",
+        "//cache/gcsproxy:go_default_library",
+        "//cache/httpproxy:go_default_library",
+        "//cache/s3proxy:go_default_library",
         "//config:go_default_library",
         "//server:go_default_library",
         "@com_github_abbot_go_http_auth//:go_default_library",
