+++ cache/gcsproxy/BUILD.bazel
@@ -2,12 +2,12 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["gcs.go"],
-    importpath = "github.com/buchgr/bazel-remote/cache/gcs",
+    srcs = ["gcsproxy.go"],
+    importpath = "github.com/buchgr/bazel-remote/cache/gcsproxy",
     visibility = ["//visibility:public"],
     deps = [
         "//cache:go_default_library",
-        "//cache/http:go_default_library",
+        "//cache/httpproxy:go_default_library",
         "@org_golang_x_oauth2//:go_default_library",
         "@org_golang_x_oauth2//google:go_default_library",
     ],
