+++ cache/httpproxy/BUILD.bazel
@@ -2,8 +2,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["http.go"],
-    importpath = "github.com/buchgr/bazel-remote/cache/http",
+    srcs = ["httpproxy.go"],
+    importpath = "github.com/buchgr/bazel-remote/cache/httpproxy",
     visibility = ["//visibility:public"],
     deps = [
         "//cache:go_default_library",
@@ -14,7 +14,7 @@ go_library(
 
 go_test(
     name = "go_default_test",
-    srcs = ["http_test.go"],
+    srcs = ["httpproxy_test.go"],
     embed = [":go_default_library"],
     deps = [
         "//cache:go_default_library",
