+++ cache/httpproxy/httpproxy_test.go
@@ -1,4 +1,4 @@
-package http
+package httpproxy
 
 import (
 	"bytes"
